+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.0.1] - 2024-04-25
+
+### Added
+
+- Add optional constructor parameter `eventStream` defaulting to false. Add
+  `.on()` and `.off()` for subscribing to pup events through WebSocketStreams if
+  eventStream is enabled.
+
+## [1.0.0] - 2024-04-23
 
 ### Added